tiny-led-firmware: Introduce mode pin
The attiny85 is too slow to handle i2c interrupts
during ws2812b update. Not handling them will block
the i2c bus and relate in bus lockups for unrelated
devices.

Fix this by introducing a mode pin (PB3), which selects
between i2c mode and led mode. If the mode pin is high,
the tiny enables the i2c interface and does not update
the leds. If the mode pin is low, the tiny updates the
leds, but does not wait for i2c updates.